% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somar.R, R/subtrair.R
\name{somar}
\alias{somar}
\title{Este programa apara somar dois numeros}
\usage{
somar(a, b)

somar(a, b)
}
\arguments{
\item{a}{um numemro}

\item{b}{outro numero}
}
\description{
Este programa apara somar dois numeros

Este programa apara somar dois numeros
}
\examples{
este pacote somar a + b

este pacote subtrais  a + b

}
